Move unmarshaledConfig to file parser.go
